Updated so GeoPath.Impl uses GeoArray.Double by default.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@2021 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/geo/GeoPath.java
@@ -113,9 +113,9 @@ public interface GeoPath extends GeoExtent {
          */
         public Impl(double[] lls, boolean isDegrees) {
             if (isDegrees) {
-                pts = GeoArray.Float.createFromLatLonDegrees(lls);
+                pts = GeoArray.Double.createFromLatLonDegrees(lls);
             } else {
-                pts = GeoArray.Float.createFromLatLonRadians(lls);
+                pts = GeoArray.Double.createFromLatLonRadians(lls);
             }
             length = pts.getSize();
         }
@@ -148,7 +148,7 @@ public interface GeoPath extends GeoExtent {
          * @param points
          */
         protected void setPoints(Geo[] points) {
-            pts = new GeoArray.Float(points);
+            pts = new GeoArray.Double(points);
             if (pts != null) {
                 length = pts.getSize();
             } else {
